[Xml] Fix code completion failure after backspacing
BXC16950 - (IntellisenseBug) Intellisense in code editor stops working
after 4th element/view in a .axml file

When backspacing or moving the caret backwards, the code completion
engine take a shortcut. Instead fo reparsing the entire file, it
clones the parser stack at a known good position before the cursor,
and reparses up to the cursor.

The problem was that cloning was not cloning the attributes. This meant
that completion for namespaced attributes failed because it did
not find the xmlns:foo on the parent element.