[Xml] Setup parent property of XElement even when the parser is not building a tree.
Since it's useful to be able to walk over the parent chain of an XElement from one of the BaseXmlEditorExtension completion methods without having to poke through the current parser Nodes collection, we setup the Parent property of each XElement right when they are created.

Previously, that property would only be set in the XContainer.AddChildNode method together with setting other attributes like sibling. This method is still called but only when the parser IXmlParserContext.BuildTree property is true.