[Xml] Fix xml validation errors not being cleared.
When closing a project the validation errors for any xml document
would not be cleared. On opening another project the status bar
would still show an error icon.

Now the xml validation errors added as tasks are associated with
the project so they are cleared when the project is closed. For
standalone files the xml text editor extension clears the error
tasks it generated when the document is closed.